create presentation models to make code testable and add heart rate info to all periods

+++ lib/past_sessions/data/model/past_sessions.dart
@@ -28,7 +28,7 @@ class SessionPeriod {
   final double beatFrequency;
   final List<BreathingPattern> breathingPattern;
   final double breathingPatternMultiplier;
-  final List<int> heartRateMeasurements;
+  final List<double> heartRateMeasurements;
   final bool isHapticFeedbackEnabled;
   final String visualization;
